Update nets, quotient maps, and Stone spaces
Introduced definitions for nets and their limits, expanded the section on quotient maps with equivalent definitions and remarks, and added propositions regarding open quotient maps and their properties in product spaces. Added the definition and equivalent characterization of Stone (profinite) spaces, as well as properties of totally disconnected spaces. Improved clarity and cross-referencing in several propositions and definitions, and made minor corrections to existing content.